% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_submap.R
\docType{data}
\name{bou_southsea}
\alias{bou_southsea}
\title{South China Sea boundary (南海十段线)}
\format{
An \code{sf} object with a single \code{MULTILINESTRING} feature (WGS 84),
covering roughly 108-123°E and 3-25°N.
}
\source{
ChinaBound2024, \code{bou1_4l_十段线.shp}.
}
\usage{
bou_southsea
}
\description{
An \code{sf} object with the "十段线" (ten-segment line) national boundary in the
South China Sea region, used as the default content of \code{\link[=geom_submap]{geom_submap()}}.
}
\seealso{
\code{\link[=geom_submap]{geom_submap()}}, \code{\link[=add_submap]{add_submap()}}
}
\keyword{datasets}
