% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar_triangle.R
\name{colourbar_triangle}
\alias{colourbar_triangle}
\title{colourbar_triangle}
\usage{
colourbar_triangle(...)
}
\arguments{
\item{...}{parameters passed to \link[ggplot2:guide_colourbar]{ggplot2::guide_colourbar}}
}
\description{
colourbar_triangle
}
\examples{
library(ggplot2)

g <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = drat))

g + scale_colour_viridis_c(
  limits = c(3, 5), oob = scales::oob_squish,
  guide = colourbar_triangle()
)
}
\references{
\enumerate{
\item \url{https://stackoverflow.com/questions/68440366/how-can-i-add-triangles-to-a-ggplot2-colorbar-in-r-to-indicate-out-of-bound-valu}
}
}
