% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar_.R
\name{convertTri}
\alias{convertTri}
\title{convertTri}
\usage{
convertTri(x, inf = FALSE, height = 0.05)
}
\arguments{
\item{x}{height of triangle. If x in the range of \verb{[0, 0.25]}, \code{height} will
be ignored.}

\item{inf}{boolean}

\item{height}{height of triangle}
}
\description{
Setting 'open.lower' and 'open.upper' to non-zero makes colorkey end with
triangular extensions, indicating open-ended intervals. Set to non-zero by
default only if first/last intervals are unbounded (-Inf / +Inf).
(NOTE: default should perhaps be 0 for back-compatibility, but currently
these are simply not shown in the legend, so probably new behaviour is no
worse).
When non-zero, controls fraction of key$height to be used for triangles at ends.
}
\keyword{internal}
