% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_lattice.R
\name{draw.colorkey}
\alias{draw.colorkey}
\alias{draw.colorkey2}
\title{draw.colorkey}
\usage{
draw.colorkey(key, draw = FALSE, vp = NULL)

draw.colorkey2(key, draw = FALSE, vp = NULL)
}
\arguments{
\item{key}{
    A list determining the key. See documentation for \code{\link[lattice]{levelplot}},
    in particular the section describing the \code{colorkey} argument,
    for details.
  }

\item{draw}{ A scalar logical, indicating whether the grob is to be drawn. }

\item{vp}{ The viewport in which to draw the grob, if applicable. }
}
\description{
draw.colorkey
}
\seealso{
\code{\link[=make_colorbar]{make_colorbar()}}
}
\keyword{internal}
