% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_polygon.R
\name{draw_polygon}
\alias{draw_polygon}
\title{draw horizontal and vertical polygons}
\usage{
draw_polygon(
  vals,
  x = NULL,
  type = "horizontal",
  length.out = 10000,
  col.regions = c("blue", "red"),
  alpha = 0.6,
  zlim = c(-Inf, Inf),
  ...
)
}
\arguments{
\item{vals}{numeric vector}

\item{x}{The corresponding x position of \code{vals}}

\item{type}{one of "horizontal" or "vertical"}

\item{length.out}{the length of interpolated \code{vals} and \code{x}}

\item{col.regions}{
    A vector of colors, or a function to produce a vecor of colors, to
    be used if \code{region=TRUE}.  Each interval defined by \code{at}
    is assigned a color, so the number of colors actually used is one
    less than the length of \code{at}.  See \code{\link[lattice]{level.colors}}
    for details on how the color assignment is done.
  }

\item{alpha}{the alpha of polygon's fill color}

\item{zlim}{limits of \code{vals}}

\item{...}{ Extra parameters. }
}
\description{
draw horizontal and vertical polygons
}
\examples{
set.seed(1)
y <- rnorm(10)
x <- seq_along(y)
}
