% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar_equispaced.R
\name{equispaced_colorkey}
\alias{equispaced_colorkey}
\title{fix non-equispaced colorkey}
\usage{
equispaced_colorkey(key)
}
\arguments{
\item{key}{A \code{colorkey} object (list), at least with the element of \code{at}.}
}
\description{
Not processed if \code{key$equispaced} is not true.
}
\examples{
x <- seq(pi/4, 5 * pi, length.out = 100)
y <- seq(pi/4, 5 * pi, length.out = 100)
r <- as.vector(sqrt(outer(x^2, y^2, "+")))
grid <- expand.grid(x=x, y=y)
grid$z <- cos(r^2) * exp(-r/(pi^3))
at = c(0.1, 0.2, 0.5, 1, Inf) \%>\% c(-rev(.), 0, .)
levelplot(z ~ x * y, grid, cuts = 50, scales=list(log="e"), xlab="",
        ylab="", main="Weird Function", sub="with log scales",
        at = at,
        colorkey = TRUE, region = TRUE) +
    theme_lattice(plot.margin = c(1, 4, 1, 0))
}
\keyword{internal}
