% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.R
\name{GOF}
\alias{GOF}
\alias{KGE}
\alias{NSE}
\title{GOF}
\usage{
GOF(obs, sim, w, include.cv = FALSE, include.r = TRUE)

KGE(obs, sim, w = c(1, 1, 1), ...)

NSE(obs, sim, w, ...)
}
\arguments{
\item{obs}{Numeric vector, observations}

\item{sim}{Numeric vector, corresponding simulated values}

\item{w}{Numeric vector, weights of every points. If w included, when
calculating mean, Bias, MAE, RMSE and NSE, w will be taken into considered.}

\item{include.cv}{If true, cv will be included.}

\item{include.r}{If true, r and R2 will be included.}

\item{...}{ignored}
}
\value{
\itemize{
\item \code{ubRMSE} unbiased root mean square error
\item \code{RMSE} root mean square error
\item \code{NSE} NASH coefficient
\item \code{MAE} mean absolute error
\item \code{AI} Agreement index (only good points (w == 1)) participate to
calculate. See details in Zhang et al., (2015).
\item \code{Bias} bias
\item \code{Bias_perc} bias percentage
\item \code{n_sim} number of valid obs
\item \code{cv} Coefficient of variation
\item \code{R2} correlation of determination
\item \code{R} pearson correlation
\item \code{pvalue} pvalue of \code{R}
}
}
\description{
Good of fitting
}
\examples{
obs = rnorm(100)
sim = obs + rnorm(100)/4
GOF(obs, sim)

}
\references{
\enumerate{
\item https://en.wikipedia.org/wiki/Coefficient_of_determination
\item https://en.wikipedia.org/wiki/Explained_sum_of_squares
\item https://en.wikipedia.org/wiki/Nash\%E2\%80\%93Sutcliffe_model_efficiency_coefficient
\item Zhang Xiaoyang (2015), http://dx.doi.org/10.1016/j.rse.2014.10.012
}
}
