% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar_add.R
\name{grobs}
\alias{grobs}
\title{grobs}
\usage{
grobs(..., options = list(nrow = 1))
}
\arguments{
\item{options}{parameters (except \code{plotlist}) of \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}}
}
\description{
grobs
}
\keyword{internal}
