% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_PosNeg.R, R/layer_PosNeg_sign.R
\name{layer_PosNeg}
\alias{layer_PosNeg}
\alias{add_PosNeg}
\alias{layer_PosNeg_sign}
\alias{add_PosNeg_sign}
\title{layer_PosNeg}
\usage{
layer_PosNeg(mapping = NULL, data = NULL, x = 0, y = 0, size = 12, ...)

add_PosNeg(
  z,
  x = 0.5,
  y = 0.5,
  height.factor = 1.2,
  cols = c("blue", "red"),
  alpha = 0.8,
  ...
)

layer_PosNeg_sign(
  mapping = NULL,
  data = NULL,
  x = 0.5,
  y = 0.5,
  height.factor = 1.2,
  size = 12,
  ...
)

add_PosNeg_sign(
  z,
  mask,
  x = 0.5,
  y = 0.5,
  height.factor = 1.2,
  cols = c("blue", "red"),
  alpha = 0.8,
  ...
)
}
\arguments{
\item{...}{other parameters to \code{\link[=element_grob_text]{element_grob_text()}}}
}
\description{
layer_PosNeg
}
