% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_latFreq.R
\name{make_latFreq}
\alias{make_latFreq}
\title{make_latFreq}
\usage{
make_latFreq(
  y,
  z,
  length.out = 10000,
  tcl = 0.4,
  cex = 1,
  xlab = "",
  ylab = "",
  xlabels = TRUE,
  ylabels = TRUE,
  ylim = NULL,
  zlim = NULL,
  zlim_ratio = c(-1, 1),
  prob_z = 0.9,
  is_spatial = FALSE,
  zticks = NULL,
  digit = 1,
  family = "Times",
  debug = FALSE,
  ...
)
}
\arguments{
\item{...}{ignored}
}
\description{
make_latFreq
}
\keyword{internal}
