% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_hatched_polygon.R
\name{polygonRingHolesLines}
\alias{polygonRingHolesLines}
\title{Get lines of one Polygons feature}
\usage{
polygonRingHolesLines(
  geom,
  density = 0.5,
  angle = 45,
  ID = 1,
  fillOddEven = FALSE
)
}
\arguments{
\item{geom}{An object of class sf POLYGON}

\item{density}{the density of shading lines, in lines per inch.  The
    default value of \code{NULL} means that no shading lines are drawn.
    A zero value of \code{density} means no shading nor filling whereas
    negative values and \code{NA} suppress shading (and so allow
    color filling).}

\item{angle}{the slope of shading lines, given as an angle
     in degrees (counter-clockwise).}

\item{ID}{Number or string identifying the Polygon inside Polygons}

\item{fillOddEven}{logical controlling the polygon shading mode:  see
      below for details.  Default \code{FALSE}.}
}
\value{
sf lines object
}
\description{
Get lines of one Polygons feature
}
\keyword{internal}
