% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_barchart.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{guide_axis_minor}
\alias{\%>\%}
\alias{\%<>\%}
\alias{mutate}
\alias{as_tibble}
\alias{tibble}
\alias{element_textbox}
\alias{element_markdown}
\alias{geom_richtext}
\alias{GeomSfPattern}
\alias{geom_sf_pattern}
\alias{slope_mk}
\alias{slope_p}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:as_tibble]{as_tibble}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:tibble]{tibble}}}

  \item{ggh4x}{\code{\link[ggh4x:guide_axis_minor]{guide_axis_minor()}}}

  \item{ggpattern}{\code{\link[ggpattern:geom_sf_pattern]{geom_sf_pattern()}}, \code{\link[ggpattern:GeomSfPattern]{GeomSfPattern}}}

  \item{ggtext}{\code{\link[ggtext:element_markdown]{element_markdown()}}, \code{\link[ggtext:element_textbox]{element_textbox()}}, \code{\link[ggtext:geom_richtext]{geom_richtext()}}}

  \item{magrittr}{\code{\link[magrittr:\%<>\%]{\%<>\%}}, \code{\link[magrittr:\%>\%]{\%>\%}}}

  \item{rtrend}{\code{\link[rtrend:slope_mk]{slope_mk()}}, \code{\link[rtrend:slope_p]{slope_p()}}}
}}

