% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_signPoint.R
\name{resample_points}
\alias{resample_points}
\title{resample_points}
\usage{
resample_points(data, fact = 1)
}
\arguments{
\item{data}{A data.frame, with the columns of "x", "y", and "mask"(optional).}

\item{fact}{postive integer. Plot one point in \code{fact} steps.}
}
\description{
resample_points
}
\keyword{internal}
