% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar_triangle.R
\name{rotate}
\alias{rotate}
\title{rotate x and y}
\usage{
rotate(x, y, rot = 0, center = c(0.5, 0.5))
}
\arguments{
\item{x, y}{numeric vector}

\item{rot}{the angle to rotate}

\item{center}{the center of rotation}
}
\description{
rotate x and y
}
\keyword{internal}
