% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar_triangle.R
\name{segmentsGrob2}
\alias{segmentsGrob2}
\title{segmentsGrob2}
\usage{
segmentsGrob2(x0, y0, x1, y1, rot = 90, ...)
}
\arguments{
\item{x0}{ Numeric indicating the starting x-values of the line segments. }

\item{y0}{ Numeric indicating the starting y-values of the line
    segments. }

\item{x1}{ Numeric indicating the stopping x-values of the line
    segments. }

\item{y1}{ Numeric indicating the stopping y-values of the line segments.}

\item{rot}{degree}
}
\description{
segmentsGrob2
}
\keyword{internal}
