% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_point2poly.R
\name{st_point2poly}
\alias{st_point2poly}
\alias{df2rast}
\alias{rast2poly}
\alias{st_dissolve}
\title{st_point2poly}
\usage{
st_point2poly(xyz, crs = 4326)

df2rast(xyz)

rast2poly(r, crs = 4326)

st_dissolve(x, by = NULL, ...)
}
\arguments{
\item{xyz}{A matrix or data frame whose first two columns are x and y;
the optional third column contains values.}

\item{crs}{A coordinate reference system accepted by \code{\link[sf:st_crs]{sf::st_crs()}}.}
}
\description{
st_point2poly
}
\references{
\enumerate{
\item https://gis.stackexchange.com/questions/192771/how-to-speed-up-raster-to-polygon-conversion-in-r
}
}
\seealso{
\code{\link[terra:rast]{terra::rast()}}, \code{\link[terra:as.polygons]{terra::as.polygons()}}
}
