% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_gof.R, R/str_format.R
\name{str_mk}
\alias{str_mk}
\title{markdown superscript and subscript}
\usage{
str_mk(x)

str_mk(x)
}
\arguments{
\item{x}{character vector}
}
\description{
markdown superscript and subscript

markdown superscript and subscript
}
\examples{
x = c(
    "gC m^{-2} d^{-1}",
    "gC m^-2 d^-1",
    "gC m_{-2} d_{-1}",
    "gC m_-2 d_-1", 
    "gC \n mm/d"
)
str_mk(x)
x <- c(
    "gC m^{-2} d^{-1}",
    "gC m^-2 d^-1",
    "gC m_{-2} d_{-1}",
    "gC m_-2 d_-1",
    "gC \n mm/d"
)
str_mk(x)
}
\author{
Dongdong Kong

Dongdong Kong
}
