% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_format.R
\name{str_num}
\alias{str_num}
\title{Rounding of Numbers, and convert to string}
\usage{
str_num(x, digits = 2)
}
\arguments{
\item{x}{a numeric vector.  Or, for \code{round} and \code{signif}, a
    complex vector.}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    Negative values are allowed (see \sQuote{Details}).}
}
\description{
Rounding of Numbers, and convert to string
}
\examples{
format(round(2, 2)) # 2
str_num(2, 2) # 2.00
}
