% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar_add.R
\name{add_colorbar}
\alias{add_colorbar}
\title{add_colorbar}
\usage{
add_colorbar(
  p,
  g,
  width = NULL,
  title = NULL,
  legend.title = element_text(hjust = 0, vjust = 0, size = 14, family = "Times")
)
}
\arguments{
\item{width}{}
}
\description{
add_colorbar
}
