% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_submap.R
\name{add_submap}
\alias{add_submap}
\title{South China Sea submap (南海小地图)}
\usage{
add_submap(
  data = NULL,
  colour = "black",
  linewidth = 0.3,
  fill = NA,
  border.colour = "black",
  border.linewidth = 0.5,
  bgcolor = "white",
  ...
)
}
\arguments{
\item{data}{An \code{sf} object of the submap boundary. If \code{NULL} (default), the
bundled \link{bou_southsea} dataset (the "十段线" national boundary in the South
China Sea) is used.}

\item{colour, linewidth, fill}{colour, line width and fill passed to
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} for the boundary line.}

\item{border.colour, border.linewidth}{colour and line width of the box drawn
around the submap (\code{panel.border}).}

\item{bgcolor}{background colour of the submap panel.}

\item{...}{other parameters passed to \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Build a small inset map of the South China Sea region, framed by a border
box. Typically placed in the bottom-right corner of a map of China via
\code{\link[=geom_submap]{geom_submap()}}.
}
\seealso{
\code{\link[=geom_submap]{geom_submap()}}, \link{bou_southsea}
}
