% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar_triangle.R
\name{colorkey_pos}
\alias{colorkey_pos}
\alias{key_gf}
\alias{key_label}
\alias{key_border}
\alias{key_triangle}
\title{\itemize{
\item \code{colorkey_pos}: The position of colorkey
}}
\usage{
colorkey_pos(space, ncol = 3)

key_gf(key, key.layout, vp, vp_label, reccentre, recdim, border = TRUE, ...)

key_label(key.gf, key, labscat, labelsGrob, vp_label, ...)

key_border(key.gf, key, open.lower, open.upper)

key_triangle(key.gf, key, open.lower, open.upper)
}
\arguments{
\item{space}{One of \code{c('top', 'bottom', 'right', 'left')}}

\item{ncol}{How many columns in right mode? This parameter is designed for
future extension.}

\item{key}{The parameter of colorkey}

\item{key.layout}{The position}

\item{vp}{viewport}

\item{vp_label}{viewport label}

\item{reccentre}{\code{reccentre = (scat[-1] + scat[-length(scat)]) / 2}}

\item{recdim}{\code{recdim = diff(scat)}}

\item{border}{boolean, the rentangle border}

\item{...}{ignored.}

\item{key.gf}{returned by \code{key.gf}}

\item{labscat}{at which to draw labels}

\item{labelsGrob}{labels grobs}

\item{open.lower, open.upper}{The width of lower and upper triangle}
}
\description{
\itemize{
\item \code{colorkey_pos}: The position of colorkey
}

key_gf_nonraster

key_label
}
\examples{
\dontrun{
key.gf <- key_gf(key, key.layout, vp, vp_label, axis.line, reccentre, recdim)
}

\dontrun{
key.gf <- key_label(key.gf, key, labscat, vp_label, axis.line)
}
}
\keyword{internal}
