% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.layers-package.R
\docType{package}
\name{gg.layers-package}
\alias{gg.layers}
\alias{gg.layers-package}
\title{gg.layers: ggplot layers}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
