% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.layers-package.R
\docType{package}
\name{gg.layers-package}
\alias{gg.layers}
\alias{gg.layers-package}
\title{gg.layers: ggplot layers}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
