% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_annotation.R, R/geom_richtext_npc.R,
%   R/guide_colorsteps2.R, R/position_nudge_npc.R, R/stat_levels.R,
%   R/stat_prob_2d.R, R/stat_signHatch.R, R/stat_signPattern.R,
%   R/stat_signPoint.R
\docType{data}
\name{GeomAnnotation}
\alias{GeomAnnotation}
\alias{GeomRichTextNpc}
\alias{GuideColoursteps2}
\alias{PositionNudgeNpc}
\alias{StatLevels}
\alias{StatPointLevels}
\alias{StatRasterLevels}
\alias{StatProb2d}
\alias{ggplot2-ggproto}
\alias{StatSignHatch}
\alias{StatSignPattern}
\alias{StatSignPoint}
\title{ggplot2-ggproto}
\description{
ggplot2-ggproto

ggplot2-ggproto
}
\details{
\code{StatLevels} is the generic, geom-agnostic engine: it bins the continuous
aesthetic \code{z} into a discrete \code{level} via \code{cut(z, breaks)}. \code{StatRasterLevels}
and \link{StatPointLevels} inherit all of this binning logic and differ only in
which aesthetic the computed \code{level} is mapped to (\code{fill} vs \code{colour}).
}
\keyword{datasets}
\keyword{internal}
