% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_signHatch.R, R/stat_signPattern.R
\docType{data}
\name{ggplot2-ggproto}
\alias{ggplot2-ggproto}
\alias{StatSignHatch}
\alias{StatSignPattern}
\title{ggplot2-ggproto}
\description{
ggplot2-ggproto
}
\keyword{datasets}
\keyword{internal}
