% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_annotation.R, R/geom_richtext_npc.R,
%   R/position_nudge_npc.R, R/stat_prob_2d.R, R/stat_signHatch.R,
%   R/stat_signPattern.R, R/stat_signPoint.R
\docType{data}
\name{GeomAnnotation}
\alias{GeomAnnotation}
\alias{GeomRichTextNpc}
\alias{PositionNudgeNpc}
\alias{StatProb2d}
\alias{ggplot2-ggproto}
\alias{StatSignHatch}
\alias{StatSignPattern}
\alias{StatSignPoint}
\title{ggplot2-ggproto}
\description{
ggplot2-ggproto

ggplot2-ggproto
}
\keyword{datasets}
\keyword{internal}
