% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colorsteps2.R
\name{guide_coloursteps2}
\alias{guide_coloursteps2}
\alias{guide_colorsteps2}
\title{guide_coloursteps2
#' @param title A title for the guide.}
\usage{
guide_coloursteps2(
  title = waiver(),
  theme = NULL,
  alpha = NA,
  even.steps = TRUE,
  show.limits = NULL,
  direction = NULL,
  reverse = FALSE,
  order = 0,
  available_aes = c("colour", "color", "fill"),
  barheight = unit(0.9, "npc"),
  ...
)

guide_colorsteps2(
  title = waiver(),
  theme = NULL,
  alpha = NA,
  even.steps = TRUE,
  show.limits = NULL,
  direction = NULL,
  reverse = FALSE,
  order = 0,
  available_aes = c("colour", "color", "fill"),
  barheight = unit(0.9, "npc"),
  ...
)
}
\arguments{
\item{theme}{A theme object for rendering the guide.}

\item{alpha}{Alpha transparency level.}

\item{even.steps}{Logical; should the steps be evenly spaced?}

\item{show.limits}{Logical; should the limits be shown?}

\item{direction}{Direction of the guide ("horizontal" or "vertical").}

\item{reverse}{Logical; should the guide be reversed?}

\item{order}{Order of the guide.}

\item{available_aes}{Available aesthetics for this guide.}

\item{barheight}{Height of the color bar. Deprecated in ggplot2 3.5.0.}

\item{...}{Additional arguments passed to the guide.
#' @example R/examples/ex-stat_cut.R}
}
\description{
guide_coloursteps2
#' @param title A title for the guide.
}
