% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colorsteps2.R
\name{guide_coloursteps2}
\alias{guide_coloursteps2}
\alias{guide_colorsteps2}
\title{guide_coloursteps2}
\usage{
guide_coloursteps2(
  title = waiver(),
  theme = NULL,
  alpha = NA,
  even.steps = TRUE,
  show.limits = NULL,
  direction = NULL,
  reverse = FALSE,
  order = 0,
  available_aes = c("colour", "color", "fill"),
  barheight = unit(0.9, "npc"),
  ...
)

guide_colorsteps2(
  title = waiver(),
  theme = NULL,
  alpha = NA,
  even.steps = TRUE,
  show.limits = NULL,
  direction = NULL,
  reverse = FALSE,
  order = 0,
  available_aes = c("colour", "color", "fill"),
  barheight = unit(0.9, "npc"),
  ...
)
}
\description{
guide_coloursteps2
}
