% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_barchart.R
\name{layer_barchart}
\alias{layer_barchart}
\title{gg_barchart}
\usage{
layer_barchart(
  mapping = NULL,
  data = NULL,
  brks,
  cols,
  x = 0,
  y = 0,
  width = unit(0.5, "npc"),
  height = unit(0.5, "npc"),
  just = c(0, 0),
  fontsize = 12,
  theme = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-location.}

\item{y}{A numeric vector or unit object specifying y-location.}

\item{width}{A numeric vector or unit object specifying width.}

\item{height}{A numeric vector or unit object specifying height.}

\item{just}{A string or numeric
    vector specifying the justification of the viewport
    relative to its (x, y) location.  If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left alignment
    and 1 means right alignment.
  }

\item{...}{Any number of grid viewport objects.}
}
\description{
gg_barchart
}
