% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar_.R
\name{make_colorbar}
\alias{make_colorbar}
\title{make_colorbar}
\usage{
make_colorbar(
  at,
  col = .regions$col,
  alpha = .regions$alpha,
  labeller = format,
  pretty = FALSE,
  equispaced = TRUE,
  tick.number = 7,
  tck = 0.2,
  tck.padding = 0,
  width = 2,
  height = 1,
  space = "right",
  raster = FALSE,
  interpolate = FALSE,
  tri.upper = NA,
  tri.lower = NA,
  title = NULL,
  unit = NULL,
  unit.adj = 0.3,
  cex.title = 1,
  axis.line = trellis.par.get("axis.line"),
  legend.text = trellis.par.get("axis.text"),
  legend.margin = bb_margin(),
  rect = list(col = "black", lwd = 0.3),
  legend.text.just = NULL,
  hjust = 0,
  vjust = 0.5,
  ...,
  draw = FALSE,
  vp = NULL
)
}
\arguments{
\item{legend.text}{\itemize{
\item \code{cex}:
\item \code{col}:
\item \code{font}:
\item \code{fontfamily}: The font family
\item \code{fontface}: The font face (bold, italic, ...)
\item \code{lineheight}:
}}

\item{legend.text.just}{The justification of the text relative to its (x, y)
location. If there are two values, the first value specifies horizontal
justification and the second value specifies vertical justification. Possible
string values are: "left", "right", "centre", "center", "bottom", and "top".
For numeric values, 0 means left (bottom) alignment and 1 means right (top)
alignment.}

\item{draw}{ A scalar logical, indicating whether the grob is to be drawn. }

\item{vp}{ The viewport in which to draw the grob, if applicable. }
}
\description{
make_colorbar
}
\examples{
brks_SOS  <- c(-Inf, seq(110, 150, 5), Inf)
key <- list(
    at = brks_SOS, 
    space = 'right', height = 1,
    tri.upper = 0.05,  tri.lower = 0.05, 
    labels=list(cex=1.2, fontface='bold'), 
    rect = list(col = "black", lwd = 0.2)
)

g <- draw.colorkey(key)
grid.newpage()
grid.draw(g)

## sp
\dontrun{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
spplot(meuse, c("ffreq"), col.regions= "black", 
       pch=c(1,2,3), key.space=list(x=0.1,y=.95,corner=c(0,1)))
spplot(meuse.grid)
}
}
