% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar.R, R/colorbar_helper.R, R/lattice.R
\name{make_colorbar}
\alias{make_colorbar}
\alias{colorkey_pos}
\alias{draw.colorkey2}
\title{make_colorbar}
\usage{
make_colorbar(
  at,
  labels = NULL,
  labeller = format,
  space = "right",
  width = 2,
  height = 1,
  col = .regions$col,
  alpha = .regions$alpha,
  pretty = FALSE,
  equispaced = TRUE,
  tick.number = 7,
  tck = 0.3,
  tck.padding = 0,
  raster = FALSE,
  interpolate = FALSE,
  tri.upper = NA,
  tri.lower = NA,
  legend.line = element_line(size = 0.8),
  legend.box = element_rect(size = 0.5),
  hjust = 0.5,
  vjust = 0.5,
  size = 12,
  family = "Times",
  legend.text.location = c(0.5, 0.5),
  legend.margin = bb_margin(),
  title = NULL,
  legend.text = element_text(hjust = 0.5),
  legend.title = element_text(),
  padding.left = unit(2, "points"),
  padding.right = unit(2, "points"),
  ...,
  draw = FALSE,
  vp = NULL
)

colorkey_pos(space = "right")

draw.colorkey2(key, draw = FALSE, vp = NULL)
}
\arguments{
\item{space}{One of \code{c('top', 'bottom', 'right', 'left')}}

\item{legend.box}{arrangement of multiple legends ("horizontal" or
"vertical")}

\item{hjust, vjust}{used in \code{\link[grid:grid.layout]{grid::grid.layout()}}}

\item{legend.margin}{the margin around each legend (\code{\link[ggplot2:element]{margin()}})}

\item{title}{all title elements: plot, axes, legends (\code{\link[ggplot2:element]{element_text()}};
inherits from \code{text})}

\item{legend.text}{\itemize{
\item \code{cex}:
\item \code{col}:
\item \code{font}:
\item \code{fontfamily}: The font family
\item \code{fontface}: The font face (bold, italic, ...)
\item \code{lineheight}:
}}

\item{legend.title}{title of legend (\code{\link[ggplot2:element]{element_text()}}; inherits from
\code{title})}

\item{padding.left, padding.right}{padding in the left and right of the legend}

\item{...}{additional element specifications not part of base ggplot2. In general,
these should also be defined in the \verb{element tree} argument.}

\item{draw}{A logical value indicating whether graphics output
    should be produced.}

\item{vp}{A Grid viewport object (or NULL).}
}
\description{
make_colorbar
}
\examples{
library(rcolors)
library(ggplot2)

brks = c(-Inf, -1, 0, 1, 3, 6, 9, Inf)
nbrk = length(brks) - 1
cols = get_color(rcolors$amwg256, nbrk)

spaces = c("right", "left", "top", "bottom") #\%>\% set_names(., .)

g <- make_colorbar(
    at = brks, col = cols, height = 1,
    tck = 0.4,
    # padding.left = unit(2, "points"),
    # padding.right = unit(2, "points"),
    space = spaces[1],
    # legend.line = element_line(size = 0.1, linetype = 1, color = "black"),
    # legend.text = element_text(hjust = 0.5),
    legend.text.location = c(0.2, 0.5),
    # legend.text.just = c(0.5, 0.5),

    title = "Title",
    legend.title = element_text(hjust = 0, vjust = 1),
    # legend.box = element_rect(size = 0.5),
    # legend.line = element_line(size = 1),
    # legend.text = list(fontfamily = "Times", cex = 1.1),
    # hjust = 0.5
)

# cowplot::plot_grid(plotlist = lst)

# Ipaper::write_fig({
#     grid.rect()
#     grid.draw(g)
# }, "a.pdf", 0.7, 4)

##
p <- ggplot(mtcars, aes(mpg, disp)) + geom_point()
p + g
# p2 = add_colorbar(p, g, width = 1)
# write_fig(p2, "a.pdf", 7, 4)
}
