% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{mutate}
\alias{GeomSfPattern}
\alias{geom_sf_pattern}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}}

  \item{ggpattern}{\code{\link[ggpattern:ggpattern-ggproto]{GeomSfPattern}}, \code{\link[ggpattern:geom-docs]{geom_sf_pattern}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}
}}

