% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_barchart.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{guide_axis_minor}
\alias{\%>\%}
\alias{\%<>\%}
\alias{mutate}
\alias{element_textbox}
\alias{element_markdown}
\alias{GeomSfPattern}
\alias{geom_sf_pattern}
\alias{slope_mk}
\alias{slope_p}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}}

  \item{ggh4x}{\code{\link[ggh4x]{guide_axis_minor}}}

  \item{ggpattern}{\code{\link[ggpattern:geom-docs]{geom_sf_pattern}}, \code{\link[ggpattern:ggpattern-ggproto]{GeomSfPattern}}}

  \item{ggtext}{\code{\link[ggtext]{element_markdown}}, \code{\link[ggtext]{element_textbox}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}

  \item{rtrend}{\code{\link[rtrend:slope]{slope_mk}}, \code{\link[rtrend:slope]{slope_p}}}
}}

