% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_hatched_polygon.R
\name{st_hatched_polygon}
\alias{st_hatched_polygon}
\title{st_hatched_polygon}
\usage{
st_hatched_polygon(x, density = 2, angle = 45, fillOddEven = FALSE)
}
\arguments{
\item{x}{sf polygon object}

\item{density}{the density of shading lines, in lines per inch. The default
value of NULL means that no shading lines are drawn. A zero value of density
means no shading nor filling whereas negative values and NA suppress shading
(and so allow color filling).}

\item{angle}{the slope of shading lines, given as an angle
     in degrees (counter-clockwise).}

\item{fillOddEven}{logical controlling the polygon shading mode:  see
      below for details.  Default \code{FALSE}.}
}
\value{
An hatched area, sf lines
}
\description{
st_hatched_polygon
}
\examples{
library(gg.layers)
library(ggplot2)
data("d_trendPerc")

d = d_trendPerc \%>\% subset(perc >= 0.99) \%>\% .[, 1:2]
poly = st_point2poly(d)
hatches = st_hatched_polygon(poly) #
ggplot(hatches) + geom_sf()
}
\references{
\enumerate{
\item \url{https://github.com/statnmap/HatchedPolygons}
\item \url{https://statnmap.github.io/HatchedPolygons/}
\item \url{https://statnmap.github.io/HatchedPolygons/articles/leaflet_shading_polygon.html}
}
}
