% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_point2poly.R
\name{st_point2poly}
\alias{st_point2poly}
\alias{df2rast}
\alias{rast2poly}
\alias{st_dissolve}
\title{st_point2poly}
\usage{
st_point2poly(xyz, crs = 4326)

df2rast(xyz)

rast2poly(r, crs = 4326)

st_dissolve(x, by = NULL, ...)
}
\arguments{
\item{xyz}{matrix or data.frame with at least three columns: x and y coordinates, and values (z). There may be several 'z' variables (columns)}

\item{crs}{one of (i) character: a string accepted by GDAL, (ii) integer, a
valid EPSG value (numeric), or (iii) an object of class crs.}
}
\description{
st_point2poly
}
\references{
\enumerate{
\item https://gis.stackexchange.com/questions/192771/how-to-speed-up-raster-to-polygon-conversion-in-r
}
}
\seealso{
\code{\link[raster:rasterFromXYZ]{raster::rasterFromXYZ()}}
}
