% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_gof.R
\name{stat_gof}
\alias{stat_gof}
\title{\code{stat} of good-of-fitting}
\usage{
stat_gof(
  mapping = NULL,
  data = NULL,
 
    format = "NSE = {str_num(NSE,2)}, R^2 = {str_num(R2, 2)} \\n RMSE = {str_num(RMSE,2)}",
  x = 0,
  y = 1,
  hjust = 0,
  vjust = 1,
  mar = 0.02,
  show.bias = TRUE
)
}
\arguments{
\item{x}{Unit objects specifying the location of the reference point.}

\item{y}{Unit objects specifying the location of the reference point.}

\item{hjust}{Numerical values specifying the justification
of the text boxes relative to \code{x} and \code{y}. These justification parameters
are specified in the internal reference frame of the text boxes, so that,
for example, \code{hjust} adjusts the vertical justification when the
text is rotated 90 degrees to the left or right.}

\item{vjust}{Numerical values specifying the justification
of the text boxes relative to \code{x} and \code{y}. These justification parameters
are specified in the internal reference frame of the text boxes, so that,
for example, \code{hjust} adjusts the vertical justification when the
text is rotated 90 degrees to the left or right.}

\item{...}{other parameters to \code{\link[gggrid:grid_panel]{gggrid::grid_panel()}}}
}
\description{
\code{stat} of good-of-fitting
}
